requirements.txt
# Data manipulation and analysis
pandas>=2.0.0
numpy>=2.2.0
scikit-learn>=1.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0

# Machine Learning and Deep Learning
xgboost>=3.0.0
prophet>=1.1.0
tensorflow>=2.12.0
keras>=2.12.0
torch>=2.7.0

# Time series analysis
statsmodels>=0.13.0

# Reinforcement Learning
stable-baselines3>=2.6.0
gymnasium>=1.1.0

# Explainable AI
shap>=0.48.0
lime>=0.2.0.1

# Web frameworks
streamlit>=1.26.0
fastapi>=0.103.1
uvicorn>=0.23.2

# API and Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.34.0
webdriver-manager>=4.0.0

# Database
sqlalchemy>=2.0.21
psycopg2-binary>=2.9.7
pymongo>=4.13.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.2
tqdm>=4.66.1
loguru>=0.7.0
scipy>=1.16.0
networkx>=3.5
sympy>=1.14.0
cloudpickle>=3.1.0

# PDF Processing
PyPDF2>=3.0.1

# Image Processing
scikit-image>=0.25.0
imageio>=2.37.0
tifffile>=2025.6.0
lazy-loader>=0.4.0

# Testing
pytest==7.4.2

# Deployment
gunicorn==21.2.0
